Moved Database Models into their own Folder called Models.
